backend_worksss/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
requests
python-dotenv
pydantic
pytest
pydantic[email]
